moved files to /methods/ subdirectory
