+++ lib/web3/allevents.js
@@ -25,7 +25,7 @@ var SolidityEvent = require('./event');
 var formatters = require('./formatters');
 var utils = require('../utils/utils');
 var Filter = require('./filter');
-var watches = require('./watches');
+var watches = require('./methods/watches');
 
 var AllSolidityEvents = function (json, address) {
     this._json = json;
