+++ lib/web3/event.js
@@ -25,7 +25,7 @@ var coder = require('../solidity/coder');
 var formatters = require('./formatters');
 var sha3 = require('../utils/sha3');
 var Filter = require('./filter');
-var watches = require('./watches');
+var watches = require('./methods/watches');
 
 /**
  * This prototype should be used to create event filters
