+++ lib/web3/methods/db.js
@@ -20,7 +20,7 @@
  * @date 2015
  */
 
-var Method = require('./method');
+var Method = require('../method');
 
 var putString = new Method({
     name: 'putString',
