+++ lib/web3/methods/watches.js
@@ -20,7 +20,7 @@
  * @date 2015
  */
 
-var Method = require('./method');
+var Method = require('../method');
 
 /// @returns an array of objects describing web3.eth.filter api methods
 var eth = function () {
