+++ test/ipcprovider.js
@@ -1,19 +1,19 @@
 var chai = require('chai');
 var assert = chai.assert;
 var SandboxedModule = require('sandboxed-module');
+var net = require('./helpers/FakeIpcRequest')
 
 SandboxedModule.registerBuiltInSourceTransformer('istanbul');
 var IpcProvider = SandboxedModule.require('../lib/web3/ipcprovider', {
     requires: {
         'bignumber.js': require('bignumber.js'), 
-        'net': require('./helpers/FakeIpcRequest')
     }
 });
 
 describe('lib/web3/ipcprovider', function () {
     describe('send', function () {
         it('should send basic request', function () {
-            var provider = new IpcProvider();
+            var provider = new IpcProvider('', net);
             var result = provider.send({id: 1, method: 'eth_test'});
 
             assert.isObject(result);
@@ -22,8 +22,7 @@ describe('lib/web3/ipcprovider', function () {
 
     describe('sendAsync', function () {
         it('should send basic async request', function (done) {
-            var provider = new IpcProvider();
-
+            var provider = new IpcProvider('', net);
             provider.sendAsync({id: 1, method: 'eth_test'}, function (err, result) {
                 assert.isObject(result);
                 done();
@@ -33,13 +32,13 @@ describe('lib/web3/ipcprovider', function () {
 
     describe('isConnected', function () {
         it('should return a boolean', function () {
-            var provider = new IpcProvider();
+            var provider = new IpcProvider('', net);
 
             assert.isBoolean(provider.isConnected());
         });
 
         it('should return false', function () {
-            var provider = new IpcProvider();
+            var provider = new IpcProvider('', net);
 
             provider.connection.writable = false;
 
@@ -47,7 +46,7 @@ describe('lib/web3/ipcprovider', function () {
         });
 
         it('should return true, when a net handle is set', function () {
-            var provider = new IpcProvider();
+            var provider = new IpcProvider('', net);
 
             provider.connection.writable = true;
